void: extract and compile code fragments
This is the raison d'être. We extend the fenced code block "info" string
to include either a filename or a shell command (preceded by bang [!]).
Coded blocks named in this way can then be extracted and used as input
for the shell command to produce the content for the block.